JollofBoy: created the header file for decode and have created the structs that will be used later on. Also created auxilliary functions to create and destroy pointers to the struct in order to keep actual memory usage low

+++ src/registers.c
@@ -54,8 +54,8 @@ uint64_t getProgramCounter(void) {
    return programCounter; 
 }
 
-// increments the program counter by n
-void incrementProgramCounter(uint64_t n) {
+// increments the program counter by n bytes 
+void incrementProgramCounter(int n) {
     programCounter += n;
 }
